Initial run of spec-kit.
## [specify.prompt.md](file:///d%3A/dev/src/github.com/kenlefeb/commute-buddy/.github/prompts/specify.prompt.md)

Build an application that will use generative AI to hold a conversation with me while I am driving to and from work, and will help me perform a Daily Review as described by the GTD methodology (https://www.gtd.be/en/how-to-get-started/maintaining-the-system/daily-reviews).

## [plan.prompt.md](file:///d%3A/dev/src/github.com/kenlefeb/commute-buddy/.github/prompts/plan.prompt.md)

The application should be built using .NET MAUI so it will run on any platform. I have expertise in .NET and AWS, but this solution should be able to be hosted in any affordable hosting platform, including my local machine.

## [tasks.prompt.md](file:///d%3A/dev/src/github.com/kenlefeb/commute-buddy/.github/prompts/tasks.prompt.md)